amend X map for AI

+++ src/GameLogic/Board.java
@@ -266,14 +266,14 @@ public class Board {
 		else if(map.equals("X")){
 		
 			//Initialise the placements of the 4 teams.
-			Point2D.Double redpos = new Point2D.Double(430, 210);
+			Point2D.Double redpos = new Point2D.Double(520, 150);
 			PhysObject red = new Square(1 ,0, 0, redpos);
 			((Square)red).setActivePlayer(true);
 			Point2D.Double blupos = new Point2D.Double(330, 210);
 			PhysObject blu = new Square(2 ,0, 0, blupos);
-			Point2D.Double yelpos = new Point2D.Double(430, 330);
+			Point2D.Double yelpos = new Point2D.Double(380, 270);
 			PhysObject yel = new Square(3 ,0, 0, yelpos);
-			Point2D.Double grnpos = new Point2D.Double(330, 330);
+			Point2D.Double grnpos = new Point2D.Double(300, 330);
 			PhysObject grn = new Square(4 ,0, 0, grnpos);
 			objects.add(red);
 			objects.add(blu);
